# swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Documentación de inventario API",
    "version": "0.0.1",
    "description": "Documentación de API utilizado para manejo de inventario"
  },
  "basePath": "/",
  "host": "localhost:3200",
  "paths": {
...
    "/api/price/product/{productId}": {
      "get": {
        "summary": "Get a price by productId",
        "description": "Get a price by productId",
        "tags": [
          "Price"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get a price by productId success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "price": {
                      "type": "object",
                      "$ref": "#/components/schemas/PriceModel"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/product": {
      "get": {
        "summary": "Get all products",
        "description": "Get all products with page and limit, optionally filtered by categories",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "tags": [
          "Product"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "description": "The page number",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "description": "How many items will be return",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "categories",
            "description": "Comma-separated category IDs to filter by (AND logic)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "list of product and total",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "products": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "$ref": "#/components/schemas/ProductModel"
                      }
                    },
                    "total": {
                      "type": "number"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a product and price",
        "description": "Create a product and price with optional image",
        "tags": [
          "Product",
          "Price"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "barcode": {
                    "type": "string",
                    "description": "Product barcode (optional)"
                  },
                  "amount": {
                    "type": "number"
                  },
                  "coin": {
                    "type": "string",
                    "$ref": "#/components/schemas/CoinType"
                  },
                  "categories": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "image": {
                    "type": "string",
                    "format": "binary",
                    "description": "Product image (jpg, jpeg, svg - max 2MB)"
                  }
                },
                "required": [
                  "name",
                  "amount",
                  "coin"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Product and Prices created success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "product": {
                      "type": "object",
                      "$ref": "#/components/schemas/ProductModel"
                    },
                    "price": {
                      "type": "object",
                      "$ref": "#/components/schemas/PriceModel"
                    }
                  },
                  "required": [
                    "product",
                    "price"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/product/barcode/{barcode}": {
      "get": {
        "summary": "Get a product by barcode",
        "description": "Get a product and its current price by barcode",
        "tags": [
          "Product"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "barcode",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Product and price",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "product": {
                      "type": "object",
                      "$ref": "#/components/schemas/ProductModel"
                    },
                    "price": {
                      "type": "object",
                      "$ref": "#/components/schemas/PriceModel"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Product not found"
          }
        }
      }
    },
    "/api/product/{productId}": {
      "get": {
        "summary": "Get a product by id",
        "description": "Get a product by id",
        "tags": [
          "Product"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Product",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "product": {
                      "type": "object",
                      "$ref": "#/components/schemas/ProductModel"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update a product",
        "description": "Update a product by id with optional image",
        "tags": [
          "Product"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "inStock": {
                    "type": "boolean"
                  },
                  "name": {
                    "type": "string"
                  },
                  "barcode": {
                    "type": "string"
                  },
                  "categories": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "image": {
                    "type": "string",
                    "format": "binary",
                    "description": "Product image (jpg, jpeg, svg - max 2MB)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Product updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "product": {
                      "type": "object",
                      "$ref": "#/components/schemas/ProductModel"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/stock/product/{productId}": {
      "get": {
        "summary": "Get stock by product ID",
        "description": "Get stock entry for a specific product",
        "tags": [
          "Stock"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get stock success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "stock": {
                      "allOf": [
                        {
                          "$ref": "#/components/schemas/StockModel"
                        },
                        {
                          "type": "object",
                          "properties": {
                            "price": {
                              "$ref": "#/components/schemas/PriceModel"
                            }
                          }
                        }
                      ]
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/stock-movement/product/{productId}": {
      "get": {
        "summary": "Get movements for a specific product",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "tags": [
          "Stock Movement"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "productId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Product ID"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 50
            },
            "description": "Max number of movements"
          }
        ],
        "responses": {
          "200": {
            "description": "Product movements"
          }
        }
      }
    },
...
    "/api/transfer/product/{productId}": {
      "get": {
        "summary": "Get stock by product across all warehouses",
        "description": "Get all stock entries for a product across all warehouses",
        "tags": [
          "Transfer"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Stock by warehouse",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "stocks": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "$ref": "#/components/schemas/StockModel"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CategoryModel": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "isEnabled": {
            "type": "boolean"
          },
          "createdBy": {
            "type": "string"
          },
          "createdAt": {
            "type": "integer"
          },
          "updatedAt": {
            "type": "integer"
          }
        },
        "required": [
          "_id",
          "name",
          "isEnabled",
          "createdBy",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "CoinType": {
        "type": "string",
        "enum": [
          "$",
          "Brs."
        ]
      },
      "PriceModel": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "amount": {
            "type": "number"
          },
          "coin": {
            "type": "string",
            "$ref": "#/components/schemas/CoinType"
          },
          "productId": {
            "type": "string"
          },
          "createdBy": {
            "type": "string"
          },
          "createdAt": {
            "type": "integer"
          },
          "updatedAt": {
            "type": "integer"
          }
        },
        "required": [
          "_id",
          "amount",
          "coin",
          "isActive",
          "productId",
          "createdBy",
          "createdAt",
          "updatedAt"
        ]
      },
      "ProductModel": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "inStock": {
            "type": "boolean"
          },
          "image": {
            "type": "string",
            "description": "Path to product image"
          },
          "barcode": {
            "type": "string",
            "description": "Product barcode (optional)"
          },
          "categories": {
            "type": "array",
            "items": {
              "type": "object",
              "$ref": "#/components/schemas/CategoryModel"
            }
          },
          "createdBy": {
            "type": "string"
          },
          "createdAt": {
            "type": "integer"
          },
          "updatedAt": {
            "type": "integer"
          }
        },
        "required": [
          "_id",
          "name",
          "createdBy",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "StockModel": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "quantity": {
            "type": "number"
          },
          "minQuantity": {
            "type": "number"
          },
          "warehouseId": {
            "type": "object",
            "properties": {
              "_id": {
                "type": "string"
              },
              "name": {
                "type": "string"
              },
              "address": {
                "type": "string"
              },
              "isEnabled": {
                "type": "boolean"
              }
            }
          },
          "productId": {
            "type": "object",
            "properties": {
              "_id": {
                "type": "string"
              },
              "name": {
                "type": "string"
              },
              "inStock": {
                "type": "boolean"
              },
              "categories": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "_id": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "isEnabled": {
                      "type": "boolean"
                    }
                  }
                }
              },
              "createdBy": {
                "type": "string"
              },
              "createdAt": {
                "type": "integer"
              },
              "updatedAt": {
                "type": "integer"
              }
            }
          },
          "createdAt": {
            "type": "integer"
          },
          "updatedAt": {
            "type": "integer"
          }
        },
        "required": [
          "_id",
          "quantity",
          "productId",
          "warehouseId",
          "createdAt",
          "updatedAt"
        ]
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your bearer token in the format **Bearer &lt;token&gt;**"
      }
    }
  },
...
}